기능구현 최종, CSS 작업중

+++ client/src/pages/ProfileInfo.tsx
@@ -7,8 +7,6 @@ import {
   MemberDetailDto,
   deleteMember,
   checkOauth2Member,
-  checkMemberPassword,
-  MemberPasswordCheckDto,
 } from "../apis/MemberApi";
 import { useState, useEffect, ChangeEvent } from "react";
 import UserInfoEditModal from "../components/modal/UserInfoEditModal";
@@ -31,10 +29,7 @@ const ProfileInfo = () => {
   const [isIntroduceEdit, setIsIntroduceEdit] = useState<boolean>(false);
   const [passowrdCheckModalOpen, setPasswordCheckModalOpen] =
     useState<boolean>(false);
-  const [passwordCheckResult, setPasswordCheckResult] =
-    useState<MemberPasswordCheckDto>({ password: "" });
   const navigate = useNavigate();
-  console.log(passwordCheckResult)
 
   // TODO 최초 페이지 진입 시 유저의 정보를 조회하는 코드
   useEffect(() => {
@@ -54,19 +49,12 @@ const ProfileInfo = () => {
   // 현재 Modal 구현은 완료했으나 비동기 처리로 인해 계속된 오류 발생. 추가적인 최적화 작업 요함
   // Jest로 테스트할 필요! : why? 소셜 로그인은 자동으로 배포 서버로 리다이렉션 함!
   const handleEditClick = async () => {
-    try {
-      const data = await checkOauth2Member(isLoggedIn);
-      if (data.provider !== "LOCAL") {
-        alert("소셜 로그인 유저는 개인정보를 수정할 수 없습니다.");
-      }
-      if (data.provider === "LOCAL") {
-        setPasswordCheckModalOpen(true);
-      }
-      checkMemberPassword(passwordCheckResult);
-    } catch (error) {
-      alert("비밀번호가 일치하지 않습니다.");
+    const data = await checkOauth2Member(isLoggedIn);
+    if (data.provider !== "LOCAL") {
+      alert("소셜 로그인 유저는 개인정보를 수정할 수 없습니다.");
+    } else {
+      setPasswordCheckModalOpen(true);
     }
-    setIsModalOpen(true);
   };
 
   // TODO Edit 버튼을 클릭 시, 유저의 자기소개, 원하는 동료상을 수정할 수 있도록 상태를 변경하는 코드
@@ -101,15 +89,19 @@ const ProfileInfo = () => {
   // TODO DELETE 버튼을 클릭 시, 유저의 자기소개 및 원하는 동료상을 서버에서 DELETE하는 코드
   const handleDeleteClick = async () => {
     try {
-      await deleteMember();
-      alert("회원탈퇴가 완료되었습니다.");
-      localStorage.clear();
-      navigate("/");
+      const confirmed = window.confirm("정말로 회원탈퇴하시겠습니까?");
+      if (confirmed) {
+        await deleteMember();
+        alert("회원탈퇴가 완료되었습니다.");
+        localStorage.clear();
+        navigate("/");
+      } else {
+        alert("회원탈퇴가 취소되었습니다.");
+      }
     } catch (error) {
       console.error(error);
     }
   };
-
   return (
     <Wrapper>
       <ProfileBaseWrapper>
@@ -123,12 +115,11 @@ const ProfileInfo = () => {
           <EditButton onClick={handleEditClick}>Edit</EditButton>
         </ProfileBaseInfo>
       </ProfileBaseWrapper>
-      {/* 유저의 자기소개와 원하는 유형의 팀원을 정리하는 자리 */}
       <IntroduceAndDesired>
         {!isIntroduceEdit ? (
           <>
-            <IntroduceAndDesiredInput value={memberInfo?.aboutMe} />
-            <IntroduceAndDesiredInput value={memberInfo?.withMe} />
+            <IntroduceAndDesiredInput value={memberInfo?.aboutMe} disabled />
+            <IntroduceAndDesiredInput value={memberInfo?.withMe} disabled />
           </>
         ) : (
           <>
@@ -144,12 +135,14 @@ const ProfileInfo = () => {
             />
           </>
         )}
-        {!isIntroduceEdit ? (
-          <EditButton onClick={handleIntroduceEditClick}>Edit</EditButton>
-        ) : (
-          <EditButton onClick={handleSaveClick}>Save</EditButton>
-        )}
-        <EditButton onClick={handleDeleteClick}>Delete</EditButton>
+        <ButtonWrapper>
+          <ExitEditButton onClick={handleDeleteClick}>회원탈퇴</ExitEditButton>
+          {!isIntroduceEdit ? (
+            <EditButton onClick={handleIntroduceEditClick}>Edit</EditButton>
+          ) : (
+            <EditButton onClick={handleSaveClick}>Save</EditButton>
+          )}
+        </ButtonWrapper>
       </IntroduceAndDesired>
       <UserInfoEditModal
         isOpen={isModalOpen}
@@ -158,7 +151,7 @@ const ProfileInfo = () => {
       <CheckPasswordModal
         isOpen={passowrdCheckModalOpen}
         closeModal={() => setPasswordCheckModalOpen(false)}
-        setPasswordCheckResult={setPasswordCheckResult}
+        setIsModalOpen={setIsModalOpen}
       />
     </Wrapper>
   );
@@ -166,11 +159,72 @@ const ProfileInfo = () => {
 
 export default ProfileInfo;
 
-const Wrapper = styled.div``;
-const ProfileBaseWrapper = styled.div``;
-const ProfileImage = styled.div``;
-const ProfileBaseInfo = styled.div``;
-const IntroduceAndDesired = styled.div``;
-const ProfileInput = styled.input``;
-const IntroduceAndDesiredInput = styled.input``;
-const EditButton = styled.button``;
+const Wrapper = styled.div`
+  width: 100%;
+  max-width: 960px;
+  margin: 0 auto;
+`;
+
+const ProfileBaseWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 20px;
+`;
+
+const ProfileImage = styled.div`
+  margin-right: 20px;
+`;
+
+const ProfileInput = styled.input`
+  margin-bottom: 10px;
+  padding: 8px;
+  width: 100%;
+`;
+
+const ProfileBaseInfo = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+`;
+
+const IntroduceAndDesired = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  margin-top: 20px;
+  width: 100%;
+`;
+
+const IntroduceAndDesiredInput = styled.input`
+  margin-bottom: 10px;
+  padding: 8px;
+  width: 100%;
+  height: 200px;
+`;
+
+const ButtonWrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+`;
+
+const EditButton = styled.button`
+  margin-bottom: 10px;
+  padding: 8px 16px;
+  background-color: #4d74b1;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+`;
+
+const ExitEditButton = styled.button`
+  margin-bottom: 10px;
+  padding: 8px 16px;
+  background-color: #7e1717;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+`;
